ai-core/requirements.txt
sentence-transformers
faiss-cpu
watchdog
numpy
